+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Additional Icon Fonts
 
+## Version 1.1.0
+
+- Fonts are now enableable individually in settings, and the module will only load enabled fonts, reducing world-load bandwidth required.
+- Only Material Design is enabled by default.
+- No longer ships with svg fonts for game-icons and jam, since as far as I can tell they were only for old iOS compatibility (shaved ~2/3 of module size).
+- Under the hood preparation for Macro & Helper Library release.
+
 ## Version 1.0.0
 
 - Add initial icon sets